Extract actions from events
Keeping them in a separate modules cuts down the size of the main
module and makes everything easier to understand.
